+++ plugin/src/cobib_dummy/__init__.py
@@ -1,8 +1,0 @@
-"""A dummy plugin for coBib.
-
-.. include:: ../../README.md
-
-.. include:: ../../../src/cobib/man/cobib-plugins.7.html_fragment
-"""
-
-__version__ = "0.0.1"